+++ core/drand_daemon.go
@@ -74,7 +74,7 @@ func NewDrandDaemon(c *Config) (*DrandDaemon, error) {
 		return nil, err
 	}
 
-	metrics.DrandStorageBackend.Set(chain.MetricsStorageType(c.dbStorageEngine))
+	metrics.DrandStorageBackend.Set(float64(chain.MetricsStorageType(c.dbStorageEngine)))
 
 	return drandDaemon, nil
 }
